+++ ReleaseNotes/release_notes.md
@@ -1,5 +1,7 @@
-
 ## New features:
+
+  1. add `operatoir()`, `density`  and `density_mass` methods for `Ostap::Kinematics::Dalitz0` and `Ostap::Kinematics::Dalitz` classes 
+
 
 ## Backward incompatible: